requirements.txt
astroid==2.2.5
attrs==19.1.0
beautifulsoup4==4.8.0
blis==0.2.4
boto3==1.9.202
botocore==1.12.202
Bottleneck==1.2.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.2
docutils==0.14
fastai==1.0.55
fastprogress==0.1.21
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonify==0.5
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
murmurhash==1.0.2
numexpr==2.6.9
numpy==1.17.0
nvidia-ml-py3==7.352.0
packaging==19.1
pandas==0.25.0
Pillow==6.1.0
plac==0.9.6
preshed==2.0.1
pylint==2.3.1
pynvx==1.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-magic==0.4.15
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.2
spacy==2.1.7
srsly==0.0.7
thinc==7.0.8
torch==1.1.0.post2
torchvision==0.3.0
tqdm==4.32.2
typed-ast==1.4.0
urllib3==1.25.3
wasabi==0.2.2
Werkzeug==0.15.5
wrapt==1.11.2
